pyproject.toml
[project]
name = "dynamic-frame-extractor"
version = "0.1.0"
description = "Extract video frames matching a text query using vision LLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.9.0",
    "openai>=1.0.0",
    "transformers>=4.57.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "accelerate>=0.26.0",
    "pillow>=10.0.0",
    "qwen-vl-utils>=0.0.14",
]

[project.scripts]
frame-extractor = "main:main"
